+++ src/xSACdb/settings/common.py
@@ -48,6 +48,9 @@ LOGIN_EXEMPT_URLS = (
     r'^accounts/',
     r'^hijack/',  # have their own protection
     r'^health/',  # Needs to be publicly accessible
+    r'^favicon.ico$',
+    r'^manifest.json$',
+    r'^service-worker.js',
 )
 
 # Absolute filesystem path to the directory that will hold user-uploaded files.
@@ -215,44 +218,6 @@ CACHES = {
     }
 }
 
-# A sample logging configuration. The only tangible logging
-# performed by this configuration is to send an email to
-# the site admins on every HTTP 500 error when DEBUG=False.
-# See http://docs.djangoproject.com/en/dev/topics/logging for
-# more details on how to customize your logging configuration.
-LOGGING = {
-    'version': 1,
-    'disable_existing_loggers': False,
-    'filters': {
-        'require_debug_false': {
-            '()': 'django.utils.log.RequireDebugFalse'
-        }
-    },
-    'handlers': {
-        'mail_admins': {
-            'level': 'ERROR',
-            'filters': ['require_debug_false'],
-            'class': 'django.utils.log.AdminEmailHandler'
-        },
-        'console': {
-            'level': 'DEBUG',
-            'class': 'logging.StreamHandler',
-        },
-    },
-    'loggers': {
-        'django.request': {
-            'handlers': ['mail_admins', 'console'],
-            'level': 'ERROR',
-            'propagate': True,
-        },
-        'django_facebook.models': {
-            'handlers': ['mail_admins', 'console'],
-            'level': 'ERROR',
-            'propagate': True,
-        }
-    }
-}
-
 AUTH_USER_MODEL = 'xsd_auth.User'
 USER_MODEL = AUTH_USER_MODEL
 AUTH_PROFILE_MODEL = 'xsd_members.MemberProfile'
@@ -295,6 +260,3 @@ ACTIVITY_MODELS = [
     ('xsd_training', 'TraineeGroup'),
     ('xsd_sites', 'Site'),
 ]
-
-## TODO HACK, disable email sending to prevent #173 on servers without email provision
-EMAIL_BACKEND = 'django.core.mail.backends.dummy.EmailBackend'