+++ src/xsd_members/urls.py
@@ -23,11 +23,13 @@ urlpatterns = patterns('',
         url(r'^member/(?P<pk>\d+)/$', MemberDetail.as_view(), name='MemberDetail'),
         url(r'^member/(?P<pk>\d+)/edit/$', MemberEdit.as_view(), name='MemberEdit'),
         url(r'^member/(?P<pk>\d+)/delete/$', MemberDelete.as_view(), name='MemberDelete'),
+        url(r'^member/(?P<pk>\d+)/archive/$', MemberArchive.as_view(), name='MemberArchive'),
 
         url(r'^list/$', MemberList.as_view(), name='MemberList'),
         url(r'^new-members/$', NewMembers.as_view(), name='NewMembers'),
         url(r'^expired-forms/$', MembersExpiredFormsList.as_view(), name='MembersExpiredFormsList'),
         url(r'^missing-fields/$', MembersMissingFieldsList.as_view(), name='MembersMissingFieldsList'),
+        url(r'^archived/$', MembersArchivedList.as_view(), name='MembersArchivedList'),
 
         url(r'^add/forms/$', BulkAddForms.as_view(), name='BulkAddForms'),